fix: issue 144

+++ api-service/internal/dto/response/tag.go
@@ -6,19 +6,19 @@ import (
 
 // TagResponse represents a tag in API responses
 type TagResponse struct {
-	ID          uint64    `json:"id" example:"1"`
+	ID          uint      `json:"id" example:"1"`
 	Name        string    `json:"name" example:"production"`
 	Color       string    `json:"color" example:"#ff0000"`
 	Description string    `json:"description" example:"Production environment tag"`
-	CreatedBy   uint64    `json:"createdBy" example:"1"`
+	CreatedBy   uint      `json:"createdBy" example:"1"`
 	CreatedAt   time.Time `json:"createdAt" example:"2024-01-01T00:00:00Z"`
 	UpdatedAt   time.Time `json:"updatedAt" example:"2024-01-01T00:00:00Z"`
-	UsageCount  int64     `json:"usageCount,omitempty" example:"25"`
+	UsageCount  int       `json:"usageCount,omitempty" example:"25"`
 }
 
 // TagSimpleResponse represents a simplified tag in API responses
 type TagSimpleResponse struct {
-	ID    uint64 `json:"id" example:"1"`
+	ID    uint   `json:"id" example:"1"`
 	Name  string `json:"name" example:"production"`
 	Color string `json:"color" example:"#ff0000"`
 }
@@ -26,7 +26,7 @@ type TagSimpleResponse struct {
 // TagAssignResult represents the result of a tag assignment operation
 type TagAssignResult struct {
 	Name    string `json:"name" example:"new-feature"`
-	TagID   uint64 `json:"tagId" example:"3"`
+	TagID   uint   `json:"tagId" example:"3"`
 	Status  string `json:"status" example:"created"`
 	Message string `json:"message" example:"Tag created and associated"`
 }
@@ -45,16 +45,16 @@ type TagUnassignResponse struct {
 
 // TaggedResource represents a resource with its associated tags
 type TaggedResource struct {
-	ResourceID   uint64              `json:"resourceId" example:"1"`
+	ResourceID   uint                `json:"resourceId" example:"1"`
 	ResourceName string              `json:"resourceName" example:"WordPress Blog"`
 	Tags         []TagSimpleResponse `json:"tags"`
-	MatchedTags  []uint64            `json:"matchedTags"`
+	MatchedTags  []uint              `json:"matchedTags"`
 	CreatedAt    time.Time           `json:"createdAt" example:"2024-01-15T10:30:00Z"`
 }
 
 // TagSearchResponse represents the response for tag search
 type TagSearchResponse struct {
-	Total     int64            `json:"total" example:"25"`
+	Total     int              `json:"total" example:"25"`
 	Page      int              `json:"page" example:"1"`
 	PageSize  int              `json:"pageSize" example:"20"`
 	Resources []TaggedResource `json:"resources"`
