+++ api-service/internal/interface/repository/tag.go
@@ -9,32 +9,32 @@ import (
 type TagRepository interface {
 	// Basic CRUD operations for tags
 	CreateTag(ctx context.Context, tag *model.Tag) error
-	GetTagByID(ctx context.Context, id uint64) (*model.Tag, error)
+	GetTagByID(ctx context.Context, id uint) (*model.Tag, error)
 	GetTagByName(ctx context.Context, name string) (*model.Tag, error)
 	UpdateTag(ctx context.Context, tag *model.Tag) error
-	DeleteTag(ctx context.Context, id uint64) error
+	DeleteTag(ctx context.Context, id uint) error
 
 	// Tag listing and searching
-	ListTags(ctx context.Context, search string, excludeIDs []uint64) ([]*model.Tag, error)
-	ListTagsWithUsageCount(ctx context.Context, search string, excludeIDs []uint64) ([]*model.Tag, error)
+	ListTags(ctx context.Context, search string, excludeIDs []uint) ([]*model.Tag, error)
+	ListTagsWithUsageCount(ctx context.Context, search string, excludeIDs []uint) ([]*model.Tag, error)
 	SearchTagsByName(ctx context.Context, query string) ([]*model.Tag, error)
 
 	// Tag existence checks
 	ExistsTagByName(ctx context.Context, name string) (bool, error)
-	ExistsTagByNameExcludeID(ctx context.Context, name string, excludeID uint64) (bool, error)
+	ExistsTagByNameExcludeID(ctx context.Context, name string, excludeID uint) (bool, error)
 
 	// Tagging operations
 	CreateTagging(ctx context.Context, tagging *model.Tagging) error
-	GetTaggingsByResourceID(ctx context.Context, resourceID uint64) ([]*model.Tagging, error)
-	GetTaggingsByTagID(ctx context.Context, tagID uint64) ([]*model.Tagging, error)
-	DeleteTagging(ctx context.Context, tagID, resourceID uint64) error
-	DeleteTaggingsByResourceID(ctx context.Context, resourceID uint64) error
-	DeleteTaggingsByTagIDs(ctx context.Context, resourceID uint64, tagIDs []uint64) error
+	GetTaggingsByResourceID(ctx context.Context, resourceID uint) ([]*model.Tagging, error)
+	GetTaggingsByTagID(ctx context.Context, tagID uint) ([]*model.Tagging, error)
+	DeleteTagging(ctx context.Context, tagID, resourceID uint) error
+	DeleteTaggingsByResourceID(ctx context.Context, resourceID uint) error
+	DeleteTaggingsByTagIDs(ctx context.Context, resourceID uint, tagIDs []uint) error
 
 	// Bulk operations
 	CreateTaggingsBatch(ctx context.Context, taggings []*model.Tagging) error
-	ExistsTagging(ctx context.Context, tagID, resourceID uint64) (bool, error)
+	ExistsTagging(ctx context.Context, tagID, resourceID uint) (bool, error)
 
 	// Search operations
-	SearchResourcesByTags(ctx context.Context, tagIDs []uint64, operation string, offset, limit int) ([]*model.Tagging, int64, error)
+	SearchResourcesByTags(ctx context.Context, tagIDs []uint, operation string, offset, limit int) ([]*model.Tagging, int, error)
 }
